requirements.txt
django
gunicorn
iommi
python-dateutil
mysqlclient
